services/notification/pyproject.toml
[tool.poetry]
name = "task-management-setup"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.2"
ipython = "^8.28.0"
mysqlclient = "^2.2.5"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
isort = "^5.13.2"
flake8 = "^7.1.1"
black = "^24.10.0"
pre-commit = "^4.0.1"
drf-spectacular = "^0.27.2"
gunicorn = "^23.0.0"
psycopg2 = "^2.9.10"
django-cors-headers = "^4.5.0"
django-soft-delete = "^1.0.15"
django-seed = "^0.3.1"
faker = "^30.8.2"
redis = "^5.2.0"
celery = "^5.4.0"
channels = "^4.1.0"
channels-redis = "^4.2.0"
django-redis = "^5.4.0"
pillow = "^11.0.0"
django-filter = "^24.3"
django-storages = "^1.14.4"
grpcio = "^1.70.0"
grpcio-tools = "^1.70.0"
googletrans = "^4.0.2"
uvicorn = "^0.34.0"
django-simple-history = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
